requirements.txt
asgiref==3.5.2
black==22.3.0
click==8.1.3
Django==4.0.4
django-cors-headers==3.4.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
djoser==2.0.3
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
postgres==4.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
psycopg2-pool==1.1
PyJWT==1.7.1
python-decouple==3.3
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2
tomli==2.0.1
